homework 22
